refactor(service-provider): 统一出海服务字段类型为Byte
- 将DTO和VO中的isAbroad字段从Integer改为Byte以匹配数据库TINYINT类型
- 更新数据库schema，为isAbroad和country_coverage字段添加NOT NULL约束
- 修正数据库设计文档中deleted字段的默认值表示方式

+++ zhilian-backend/src/main/java/com/zhilian/zhilianbackend/dto/request/ServiceProviderUpdateRequestDTO.java
@@ -61,7 +61,7 @@ public class ServiceProviderUpdateRequestDTO {
     private String qualification;
 
     @Schema(description = "是否提供出海服务（0否 1是）", example = "1")
-    private Integer isAbroad;
+    private Byte isAbroad;
 
     @Schema(description = "覆盖国家/地区，多个用逗号分隔", example = "欧盟,美国")
     private String countryCoverage;
